+++ src/test/psdReader.spec.ts
@@ -68,7 +68,7 @@ describe('PsdReader', () => {
 	});
 
 	fs.readdirSync(readFilesPath).filter(f => !/pattern/.test(f)).forEach(f => {
-	// fs.readdirSync(readFilesPath).filter(f => /text-simple2/.test(f)).forEach(f => {
+		// fs.readdirSync(readFilesPath).filter(f => /text-simple2/.test(f)).forEach(f => {
 		it(`reads PSD file (${f})`, () => {
 			const basePath = path.join(readFilesPath, f);
 			const fileName = fs.existsSync(path.join(basePath, 'src.psb')) ? 'src.psb' : 'src.psd';
@@ -138,7 +138,7 @@ describe('PsdReader', () => {
 	});
 
 	fs.readdirSync(readWriteFilesPath).forEach(f => {
-		// fs.readdirSync(readWriteFilesPath).filter(f => /vectors/.test(f)).forEach(f => {
+		// fs.readdirSync(readWriteFilesPath).filter(f => /werps/.test(f)).forEach(f => {
 		it(`reads-writes PSD file (${f})`, () => {
 			const ext = fs.existsSync(path.join(readWriteFilesPath, f, 'src.psb')) ? 'psb' : 'psd';
 			const psd = readPsdFromFile(path.join(readWriteFilesPath, f, `src.${ext}`), {
@@ -147,13 +147,14 @@ describe('PsdReader', () => {
 			const actual = writePsdBuffer(psd, { logMissingFeatures: true, psb: ext === 'psb' });
 			const expected = fs.readFileSync(path.join(readWriteFilesPath, f, `expected.${ext}`));
 			fs.writeFileSync(path.join(resultsFilesPath, `read-write-${f}.${ext}`), actual);
-			// console.log(require('util').inspect(psd, false, 99, true));
-			// fs.writeFileSync('temp.txt', require('util').inspect(psd, false, 99, false), 'utf8');
 			fs.writeFileSync(path.join(resultsFilesPath, `read-write-${f}.bin`), actual);
-			// console.log('------------');
-			// readPsdFromFile(path.join(resultsFilesPath, `read-write-${f}.psd`), { ...opts, useImageData: true, useRawThumbnail: true });
+			// console.log(require('util').inspect(psd, false, 99, true));
 
-			compareBuffers(actual, expected, `read-write-${f}`, 0);
+			// const psd2 = readPsdFromFile(path.join(resultsFilesPath, `read-write-${f}.psd`), { ...opts, useImageData: true, useRawThumbnail: true });
+			// fs.writeFileSync('temp.txt', require('util').inspect(psd, false, 99, false), 'utf8');
+			// fs.writeFileSync('temp2.txt', require('util').inspect(psd2, false, 99, false), 'utf8');
+
+			compareBuffers(actual, expected, `read-write-${f}`, 0x81e10);
 		});
 	});